Gate usable_size behind extended feature
